update: convert state management to redux toolkit structure
