+++ frontend/src/App.js
@@ -30,6 +30,6 @@ function App() {
       </Router>
     </div>
   );
-}
+};
 
 export default App;
