+++ frontend/src/index.js
@@ -1,17 +1,14 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import  { Provider } from 'react-redux'; 
-import { PersistGate } from 'redux-persist/integration/react';
+import { Provider } from 'react-redux'; 
 
-import { store, persistor } from './redux/store';
+import store from './app/store';
 import App from './App';
 
 ReactDOM.render(
   <React.StrictMode>
     <Provider store={store}>
-      <PersistGate persistor={persistor}>
         <App />
-      </PersistGate>
     </Provider>
   </React.StrictMode>,
   document.getElementById('root')
