--- frontend/src/redux/actions/userActions.js
@@ -1,15 +1,0 @@
-import { ActionTypes } from "../constants/action-constants";
-
-export const getResource = (resource) => {
-  return {
-    type: ActionTypes.GET_ALL,
-    payload: resource
-  };
-};
-
-export const getFollowings = (following) => {
-  return {
-    type: ActionTypes.GET_FOLLOWINGS,
-    payload: following
-  };
-};
