--- frontend/src/redux/constants/action-constants.js
@@ -1,5 +1,0 @@
-export const ActionTypes = {
-  GET_ALL: 'GET_ALL',
-
-  GET_FOLLOWINGS: 'GET_FOLLOWINGS',
-};
